+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
+- Added: `CCSave.fromObject` automatically sets `building.highest` if absent.
 - Fixed: `setupCookieClickerPage` was not exported.
 
 ## 0.1.3 - 2021-09-13